+++ bean-searcher/src/test/java/cn/zhxu/bs/convertor/BoolParamConvertorTestCase.java
@@ -28,7 +28,7 @@ public class BoolParamConvertorTestCase {
     }
 
     void assertSupports(DbType dbType, boolean supports) {
-        final FieldMeta meta = new FieldMeta(null, null, null, null, false, null, dbType);
+        final FieldMeta meta = new FieldMeta(null, null,null, null, null, false, null, dbType);
         Assert.assertEquals(supports, convertor.supports(meta, Byte.class));
         Assert.assertEquals(supports, convertor.supports(meta, Short.class));
         Assert.assertEquals(supports, convertor.supports(meta, Integer.class));
@@ -39,7 +39,7 @@ public class BoolParamConvertorTestCase {
 
     @Test
     public void test_convert() {
-        FieldMeta meta = new FieldMeta(null, null, null, null, false, null, DbType.BOOL);
+        FieldMeta meta = new FieldMeta(null, null,null, null, null, false, null, DbType.BOOL);
         Assert.assertEquals(true, convertor.convert(meta, 1));
         Assert.assertEquals(true, convertor.convert(meta, true));
         Assert.assertEquals(true, convertor.convert(meta, "true"));