+++ bean-searcher/src/test/java/cn/zhxu/bs/convertor/DateTimeParamConvertorTestCase.java
@@ -36,7 +36,7 @@ public class DateTimeParamConvertorTestCase {
     }
 
     void assertSupports(DbType dbType, boolean supports) {
-        FieldMeta meta = new FieldMeta(null, null, null, null, false, null, dbType);
+        FieldMeta meta = new FieldMeta(null,  null, null, null, null, false, null, dbType);
         Assert.assertEquals(supports, convertor.supports(meta, Date.class));
         Assert.assertEquals(supports, convertor.supports(meta, LocalDate.class));
         Assert.assertEquals(supports, convertor.supports(meta, LocalDateTime.class));
@@ -63,7 +63,7 @@ public class DateTimeParamConvertorTestCase {
     }
 
     private void assertConvert(Object value, int hour, int minutes, int seconds) {
-        FieldMeta meta = new FieldMeta(null, null, null, null, false, null, DbType.DATETIME);
+        FieldMeta meta = new FieldMeta(null, null,null, null, null, false, null, DbType.DATETIME);
         Object date = convertor.convert(meta, value);
         Assert.assertTrue(date instanceof java.sql.Timestamp);
         Calendar calendar = Calendar.getInstance();