+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   * 功能：注解 `@SearchBean` 新增 `fields` 属性，用于指定额外的动态条件字段
   * 优化：注解 `@SearchBean.groupBy` 使用拼接参数时，条件字段的生成逻辑
   * 优化：完善 `groupBy` 与 `groupExpr` 同时使用时 `where` 与 `having` 的条件拆分逻辑
+  * 优化：增强逻辑表达式的自动化简能力
 * Bean Searcher Boot Starter
   * GroupPairResolver 支持注入 Bean 的方式自定义
 * Bean Searcher Solon Plugin